minor refactor in some components
