+++ lib/browser_core.js
@@ -94,7 +94,8 @@ module.exports = class BrowserCore {
 
     _beforeClose() {
         return Promise.all(this._components.map((c) => {
-            return this[c]._beforeClose();
+            if (typeof this[c]._beforeClose === 'function')
+                return this[c]._beforeClose();
         }));
     }
 
@@ -106,7 +107,8 @@ module.exports = class BrowserCore {
         return p.then(() => {
             return this.setViewport(options.viewport).then(() => {
                 return Promise.all(this._components.map((c) => {
-                    return this[c]._beforeOpen(options);
+                    if (typeof this[c]._beforeOpen === 'function')
+                        return this[c]._beforeOpen(options);
                 }));
             });
         });
