+++ lib/modules/assertions/browser_assertions.js
@@ -3,17 +3,15 @@
 "use strict";
 
 const utils = require('../../utils');
-const elementsAssertionUtils = require('./utils/assert_elements');
-const assertUtils = require('./utils/assert_utils');
-const BrowserModule = require('../browser_module');
+const elementsAssertionUtils = require('../../utils/assert_elements');
+const assertUtils = require('../../utils/assert_utils');
 const RequestAssertionsFilter = require('./request_assertions_filter');
 const BrowserNotAssertions = require('./browser_not_assertions');
 const {QueryError} = require('../../errors');
 
-module.exports = class BrowserAssertions extends BrowserModule {
+module.exports = class BrowserAssertions {
     constructor(browser) {
-        super(browser);
-
+        this._browser = browser;
         this.not = new BrowserNotAssertions(browser, this);
     }
