+++ lib/modules/assertions/browser_not_assertions.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const utils = require('../../utils');
-const assertUtils = require('./utils/assert_utils');
+const assertUtils = require('../../utils/assert_utils');
 
 module.exports = class BrowserNotAssertions {
     constructor(browser, browserAssertions) {
