+++ lib/modules/browser_cookies.js
@@ -1,8 +1,10 @@
 "use strict";
 
-const BrowserModule = require('./browser_module');
+module.exports = class BrowserCookies {
+    constructor(browser) {
+        this._browser = browser;
+    }
 
-module.exports = class BrowserCookies extends BrowserModule {
     all() {
         return this._browser.page.cookies().then((cookies) => {
             return cookies.reduce((acc, cookie) => {
