--- lib/modules/browser_module.js
@@ -1,15 +1,0 @@
-"use strict";
-
-module.exports = class BrowserModule {
-    constructor(browser) {
-        this._browser = browser;
-    }
-
-    _beforeOpen() {
-        // Override only if need to perform actions on open
-    }
-
-    _beforeClose() {
-        // Override only if need to perform actions on open
-    }
-};
