+++ lib/modules/browser_requests.js
@@ -2,11 +2,10 @@
 
 const RequestFilter = require('../request_filter');
 const RequestMocker = require('../request_mocker');
-const BrowserModule = require('./browser_module');
 
-module.exports = class BrowserRequests extends BrowserModule {
+module.exports = class BrowserRequests {
     constructor(browser) {
-        super(browser);
+        this._browser = browser;
         this._requestMocker = new RequestMocker();
         this._requests = [];
         this._interceptorReady = false;
