+++ lib/modules/browser_webworkers.js
@@ -1,9 +1,12 @@
 "use strict";
 
-const BrowserModule = require('./browser_module');
 const WebWorker = require('../models/webworker');
 
-module.exports = class BrowserWebWorkers extends BrowserModule {
+module.exports = class BrowserWebWorkers {
+    constructor(browser) {
+        this._browser = browser;
+    }
+
     all() {
         return this._browser.page.workers().map((ww) => {
             return new WebWorker(ww);
