+++ lib/modules/console/browser_console.js
@@ -1,7 +1,6 @@
 "use strict";
 
-const BrowserModule = require('./browser_module');
-const utils = require('../utils');
+const utils = require('../../utils');
 
 const LogType = {
     log: 'log',
@@ -26,9 +25,9 @@ class Log {
     }
 }
 
-module.exports = class BrowserConsole extends BrowserModule {
+module.exports = class BrowserConsole {
     constructor(browser) {
-        super(browser);
+        this._browser = browser;
         this.clear();
         this._browser.page.on("console", (log) => {
             utils.stringifyLogText(log).then(text => {
