+++ lib/modules/local_storage/browser_local_storage.js
@@ -1,8 +1,10 @@
 "use strict";
 
-const BrowserModule = require('./browser_module');
+module.exports = class BrowserLocalStorage {
+    constructor(browser) {
+        this._browser = browser;
+    }
 
-module.exports = class BrowserLocalStorage extends BrowserModule {
     getItem(key) {
         return this._browser.evaluate((k) => {
             return localStorage.getItem(k);
