+++ lib/modules/local_storage/local_storage_assertions.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const BrowserLocalStorageNotAssertions = require('./local_storage_not_assertions');
-const assertUtils = require('./utils/assert_utils');
+const assertUtils = require('../../utils/assert_utils');
 
 module.exports = class BrowserLocalStorageAssertions {
     constructor(browser, localStorage) {
