+++ lib/modules/local_storage/local_storage_not_assertions.js
@@ -1,6 +1,7 @@
 "use strict";
 
-const assertUtils = require('./utils/assert_utils');
+const assertUtils = require('../../utils/assert_utils');
+
 
 module.exports = class BrowserLocalStorageNotAssertions {
     constructor(localStorageAssertions) {
