+++ lib/utils/assert_elements.js
@@ -1,7 +1,7 @@
 "use strict";
 
 const assertUtils = require('./assert_utils');
-const utils = require('../../../utils');
+const utils = require('../utils');
 
 const countCases = {
     equal: "equal",
