+++ lib/utils/assert_utils.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const {AssertionError} = require('../../../errors');
+const {AssertionError} = require('../errors');
 
 
 module.exports = {
