+++ lib/wendigo.js
@@ -21,14 +21,14 @@ const defaultPlugins = [{
     plugin: require('./modules/browser_cookies')
 }, {
     name: "localStorage",
-    plugin: require('./modules/browser_local_storage'),
-    assertions: require('./modules/assertions/local_storage_assertions')
+    plugin: require('./modules/local_storage/browser_local_storage'),
+    assertions: require('./modules/local_storage/local_storage_assertions')
 }, {
     name: "requests",
     plugin: require('./modules/browser_requests')
 }, {
     name: "console",
-    plugin: require('./modules/browser_console')
+    plugin: require('./modules/console/browser_console')
 }, {
     name: "webworkers",
     plugin: require('./modules/browser_webworkers')
